W14T5: Do not use CNAMEs for AGAMA MySQL
It's not clear from the task does `db-1` points to vm-1 or MySQL master. The latter will break the switchover (lab 11).

It's enough to create a few CNAMEs in 1..2 roles to understand how it works, covering all infra with CNAMEs is not justified IMHO.

Also `lb-N` pints to the same machine as `www-N`, we probably don't need both.

Fixed https://github.com/ica0002-bot/ansible/issues/118.